fixing remaining issues
